+++ src/lib.rs
@@ -52,6 +52,7 @@ mod develop;
 mod metadata;
 mod module_writer;
 mod new_project;
+mod project_layout;
 mod pyproject_toml;
 mod python_interpreter;
 mod source_distribution;